+++ modules/base.nix
@@ -194,7 +194,8 @@ in {
       # https://nix.dev/anti-patterns/language#with-attrset-expression
       home.packages = builtins.attrValues {
         inherit (pkgs)
-          peco ripgrep jq htop ranger tmux tree magic-wormhole-rs hishtory;
+          peco ripgrep jq htop ranger tmux tree magic-wormhole-rs hishtory
+          zellij;
 
         reptyr = lib.mkIf hostPlatform.isLinux pkgs.reptyr;
       };