+++ modules/terranix/base.nix
@@ -2,12 +2,44 @@
 
 let clan = inputs'.clan-core.packages.clan-cli;
 in {
+  terraform.backend.s3 = {
+    endpoints.s3 = "https://s3.us-west-001.backblazeb2.com";
+    bucket = "enzime-dotfiles-tf-state";
+    key = "tofu.tfstate";
+    region = "us-west-001";
+
+    skip_credentials_validation = true;
+    skip_region_validation = true;
+    skip_metadata_api_check = true;
+    skip_requesting_account_id = true;
+    skip_s3_checksum = true;
+
+    skip_bucket_root_access = true;
+    skip_bucket_enforced_tls = true;
+  };
+
+  terraform.encryption = {
+    key_provider.external.passphrase = {
+      command = [ (lib.getExe self'.packages.provide-tf-passphrase) ];
+    };
+
+    key_provider.pbkdf2.state_encryption_password = {
+      chain = lib.tf.ref "key_provider.external.passphrase";
+    };
+
+    method.aes_gcm.encryption_method.keys =
+      lib.tf.ref "key_provider.pbkdf2.state_encryption_password";
+
+    state.enforced = true;
+    state.method = "method.aes_gcm.encryption_method";
+
+    plan.enforced = true;
+    plan.method = "method.aes_gcm.encryption_method";
+  };
+
   terraform.required_providers.local.source = "hashicorp/local";
-  terraform.required_providers.onepassword.source = "1Password/onepassword";
   terraform.required_providers.tailscale.source = "tailscale/tailscale";
   terraform.required_providers.tls.source = "hashicorp/tls";
-
-  provider.onepassword.account = "my.1password.com";
 
   data.external.tailscale-api-key = {
     program =