+++ overlays/firefox-addons/flake.nix
@@ -24,6 +24,16 @@
             throw "firefox-addons.${name} already exists"
           else
             addon) {
+              masked-email-manager = buildFirefoxXpiAddon {
+                pname = "masked-email-manager";
+                version = "1.7.2";
+                addonId = "{c48d361c-1173-11ee-be56-0242ac120002}";
+                url =
+                  "https://addons.mozilla.org/firefox/downloads/file/4585287/masked_email_manager-1.7.2.xpi";
+                sha256 = "sha256-UBcHS4ackk1RpWTRAyj8SB2VYe4hVIfive23hN/hd1I=";
+                meta = { };
+              };
+
               purple-private-windows = buildFirefoxXpiAddon {
                 pname = "purple-private-windows";
                 version = "1.1";